refactor(sessions): dedupe session helpers
