+++ src/config/sessions.ts
@@ -1,5 +1,6 @@
 // Public facade for session stores, metadata, lifecycle, reset, transcript, and cleanup APIs.
 export * from "./sessions/combined-store-gateway.js";
+export * from "./sessions/compaction-session-file.js";
 export * from "./sessions/group.js";
 export * from "./sessions/goals.js";
 export * from "./sessions/artifacts.js";
