+++ src/config/sessions/goals.ts
@@ -1,5 +1,6 @@
 // Session goal state tracks objective progress and token budgets in the session store.
 import crypto from "node:crypto";
+import { formatTokenCount } from "../../utils/token-format.js";
 import { getSessionEntry, patchSessionEntry } from "./store.js";
 import { resolveFreshSessionTotalTokens } from "./types.js";
 import type { SessionEntry, SessionGoal, SessionGoalStatus } from "./types.js";
@@ -60,25 +61,6 @@ function normalizeTokenBudget(value: number | undefined): number | undefined {
 
 function cloneGoal(goal: SessionGoal): SessionGoal {
   return { ...goal };
-}
-
-function formatGoalTokenCount(value: number | undefined): string {
-  if (value === undefined || !Number.isFinite(value)) {
-    return "0";
-  }
-  const safe = Math.max(0, value);
-  if (safe >= 1_000_000) {
-    return `${(safe / 1_000_000).toFixed(1)}m`;
-  }
-  if (safe >= 1_000) {
-    const precision = safe >= 10_000 ? 0 : 1;
-    const formattedThousands = (safe / 1_000).toFixed(precision);
-    if (Number(formattedThousands) >= 1_000) {
-      return `${(safe / 1_000_000).toFixed(1)}m`;
-    }
-    return `${formattedThousands}k`;
-  }
-  return String(Math.round(safe));
 }
 
 export function resolveSessionGoalDisplayState(
@@ -143,14 +125,14 @@ export function formatSessionGoalStatus(goal: SessionGoal | undefined): string {
   const budget =
     goal.tokenBudget === undefined
       ? ""
-      : `\nToken budget: ${formatGoalTokenCount(goal.tokensUsed)}/${formatGoalTokenCount(goal.tokenBudget)}`;
+      : `\nToken budget: ${formatTokenCount(goal.tokensUsed)}/${formatTokenCount(goal.tokenBudget)}`;
   const note = goal.lastStatusNote ? `\nNote: ${goal.lastStatusNote}` : "";
   const commands = resolveGoalCommandHint(goal.status);
   return [
     "Goal",
     `Status: ${goal.status}`,
     `Objective: ${goal.objective}`,
-    `Tokens used: ${formatGoalTokenCount(goal.tokensUsed)}`,
+    `Tokens used: ${formatTokenCount(goal.tokensUsed)}`,
     ...(budget ? [budget.slice(1)] : []),
     ...(note ? [note.slice(1)] : []),
     "",
